[HOPSWORKS-3223] feature view refactoring (#1003)
