+++ hopsworks-api/src/main/java/io/hops/hopsworks/api/featurestore/activities/ActivityResource.java
@@ -16,7 +16,7 @@
 
 package io.hops.hopsworks.api.featurestore.activities;
 
-import io.hops.hopsworks.api.featurestore.featureview.FeatureViewController;
+import io.hops.hopsworks.common.featurestore.featureview.FeatureViewController;
 import io.hops.hopsworks.api.filter.AllowedProjectRoles;
 import io.hops.hopsworks.api.filter.Audience;
 import io.hops.hopsworks.api.filter.apiKey.ApiKeyRequired;
