+++ hopsworks-api/src/main/java/io/hops/hopsworks/api/featurestore/featureview/FeatureViewInputValidator.java
@@ -16,7 +16,6 @@
 
 package io.hops.hopsworks.api.featurestore.featureview;
 
-import io.hops.hopsworks.common.featurestore.featureview.FeatureViewDTO;
 import io.hops.hopsworks.common.featurestore.query.Query;
 import io.hops.hopsworks.common.featurestore.query.QueryController;
 import io.hops.hopsworks.common.featurestore.trainingdatasets.TrainingDatasetInputValidation;
