+++ hopsworks-api/src/main/java/io/hops/hopsworks/api/featurestore/featureview/FeatureViewResource.java
@@ -23,7 +23,7 @@ import io.hops.hopsworks.api.filter.apiKey.ApiKeyRequired;
 import io.hops.hopsworks.api.jwt.JWTHelper;
 import io.hops.hopsworks.common.api.ResourceRequest;
 import io.hops.hopsworks.common.dao.QueryParam;
-import io.hops.hopsworks.common.featurestore.featureview.FeatureViewDTO;
+import io.hops.hopsworks.common.featurestore.featureview.FeatureViewController;
 import io.hops.hopsworks.exceptions.DatasetException;
 import io.hops.hopsworks.exceptions.FeaturestoreException;
 import io.hops.hopsworks.exceptions.MetadataException;
@@ -105,7 +105,7 @@ public class FeatureViewResource {
     }
     Users user = jWTHelper.getUserPrincipal(sc);
 
-    FeatureView featureView = featureViewController.convertFromDTO(project, featurestore, user, featureViewDTO);
+    FeatureView featureView = featureViewBuilder.convertFromDTO(project, featurestore, user, featureViewDTO);
     featureView = featureViewController.createFeatureView(project, user, featureView, featurestore);
 
     ResourceRequest resourceRequest = new ResourceRequest(ResourceRequest.Name.FEATUREVIEW);
@@ -296,7 +296,8 @@ public class FeatureViewResource {
     }
     Users user = jWTHelper.getUserPrincipal(sc);
     ResourceRequest resourceRequest = makeResourceRequest(param);
-    FeatureView featureView = featureViewController.update(user, project, featurestore, name, version, featureViewDTO);
+    FeatureView featureView = featureViewController.update(user, project, featurestore, name, version,
+      featureViewDTO.getDescription());
 
     return Response.ok()
         .entity(featureViewBuilder.build(featureView, resourceRequest, project, user, uriInfo))
