+++ hopsworks-api/src/main/java/io/hops/hopsworks/api/featurestore/query/QueryResource.java
@@ -17,7 +17,7 @@
 package io.hops.hopsworks.api.featurestore.query;
 
 import io.hops.hopsworks.api.featurestore.FsQueryBuilder;
-import io.hops.hopsworks.api.featurestore.featureview.FeatureViewController;
+import io.hops.hopsworks.common.featurestore.featureview.FeatureViewController;
 import io.hops.hopsworks.api.filter.AllowedProjectRoles;
 import io.hops.hopsworks.api.filter.Audience;
 import io.hops.hopsworks.api.filter.apiKey.ApiKeyRequired;
